Support typesetting by both latex and pdflatex
This change ensures that the correct output drivers are loaded for the selected processing workflow. Loading the DVI drivers for latex manually is necessary because DVI output can't be auto-detected. Limiting this with \Ifdvioutput means that either of the latex or pdflatex tools can be used for typesetting runs without changes to the LaTeX source code.

That said, latex is curently preferred over pdflatex because DVI output leads to smaller PDF file sizes.