Add Carlsbad NM
A truck wash is not currently a usable facility or feature of interest, but it seems this is the best way to describe this particular rest area. The alternative, which is to label it as "rest area" and explain in the description that it's behind a truck wash, is less concise.