Update for 1.51: Game economy gas station brands
Version 1.51 brings most of the in-world gas station brands to the game economy, making them visible in the Job Chooser and other tools.

Henry's Ranch is now the only gas station brand that intentionally uses an underground tank of a different company (it's still treated as a Vortex location by the economy). This is understandable, as there's only a single such location.

There are still a few bugs and questionable cases, namely in Garden City KS, Huntsville TX, Kalispell MT, Laredo TX, Lufkin TX. The C/FD tries to describe the current situation accurately in an effort to account for these issues.

The location in Glendive MT probably changed its in-world brand in 1.50 already, but the economy was only changed in 1.51.