Complete Oregon based on driving directions
This is an attempt to get some quick-and-dirty content for Oregon by recycling the original driving directions I wrote as a proof-of-concept back in early 2019. The quality of the result varies from place to place. My impression is that it works surprisingly well in many cases.

However, the conversion of the long-winded directions to brevity-focused location decriptions inevitably results in the loss of many details. Some of those details may have been important. A new field-check would be required to tell. Additionally, I know that SCS made improvements to the signage in Oregon. Taking advantage of those here would require a field-check as well.

The 2019 driving directions didn't cover trucking facilities such as gas stations. These had to be added for this work. I used Trucker Mudgeon's map for that.

The recorded time only covers the conversion itself, not the original research in 2019.