Allow cross-chapter city references with fallback text
Depending on how the document is built, references to other chapters may not be defined. In particular, clean runs on single chapters can be a problem.

This change allows \CityRef to take an optional fallback value to be used as reference in case the given city's label doesn't exist in the aux files. When fallbacks are given, no warnings for undefined references will appear during single-chapter runs.

The fallback isn't meant to be a perfect replacement for the real reference. Giving exact page numbers isn't useful, as they would have to be maintained manually. A pointer to the right chapter should generally be enough.