Complete Nevada, remove supplement
Some "quick and dirty" work is probably better than nothing at all. I didn't really plan on writing Nevada at this time, but ended up doing so anyway just to get it out of my head.

Cities in northern Nevada show up as "field-check needed" on the status map. Semantically, that's not entirely accurate. I did check it in-game, but the lack of proper signage just makes it very difficult to describe these locations well and concisely. The effect is the same though: Drivers headed to these cities will have to be a little bit careful. Thus "field-check needed" is actually correct on the pragmatic level.

Since there are no incomplete states left, there is no longer a need for a separate supplemental document.