Reformat files for stats and overview map
The filenames should only use "portable" chars according to POSIX. (The previous names were from my own local use on macOS, which allows spaces and / in filenames, with / being coded as : on the file system level.)

In the stats table, the count and time columns were useful to calculate the progress in terms of catching up with SCS (see the old wiki help page for details). They shouldn't be needed anymore.

The center of the status overview map should ideally be north-aligned.

The link target of the readme map image should be the table, so as to offer additional details. Image and link should refer to the current branch, not the main branch.