Rework incomplete chapter declaration and disclaimers
Since there is no supplement anymore, partial work on new states should probably be available in the full document instead. Such incomplete chapters need a new inline disclaimer in the full document.

The declaration of a state as incomplete is moved to an optional parameter of the chapter command, which should be more reliable.